Revert 64bit requirement
There is no reason for it, the timestamps can be float with no precision loss until October 12, year 287396.

+++ lib/Db/ItemMapperV2.php
@@ -324,7 +324,7 @@ class ItemMapperV2 extends NewsMapperV2
     /**
      * @param string $userId
      * @param int    $feedId
-     * @param int    $updatedSince
+     * @param float  $updatedSince
      * @param bool   $hideRead
      *
      * @return Item[]
@@ -332,7 +332,7 @@ class ItemMapperV2 extends NewsMapperV2
     public function findAllInFeedAfter(
         string $userId,
         int $feedId,
-        int $updatedSince,
+        float $updatedSince,
         bool $hideRead
     ): array {
         $builder = $this->db->getQueryBuilder();
@@ -362,7 +362,7 @@ class ItemMapperV2 extends NewsMapperV2
     /**
      * @param string   $userId
      * @param int|null $folderId
-     * @param int      $updatedSince
+     * @param float    $updatedSince
      * @param bool     $hideRead
      *
      * @return Item[]
@@ -370,7 +370,7 @@ class ItemMapperV2 extends NewsMapperV2
     public function findAllInFolderAfter(
         string $userId,
         ?int $folderId,
-        int $updatedSince,
+        float $updatedSince,
         bool $hideRead
     ): array {
         $builder = $this->db->getQueryBuilder();
@@ -396,13 +396,13 @@ class ItemMapperV2 extends NewsMapperV2
 
     /**
      * @param string $userId
-     * @param int    $updatedSince
+     * @param float  $updatedSince
      * @param int    $feedType
      *
      * @return Item[]|Entity[]
      * @throws ServiceValidationException
      */
-    public function findAllAfter(string $userId, int $feedType, int $updatedSince): array
+    public function findAllAfter(string $userId, int $feedType, float $updatedSince): array
     {
         $builder = $this->db->getQueryBuilder();
